refactor:페이지응답통일

+++ src/main/java/com/threestar/trainus/domain/coupon/admin/dto/CouponListResponseDto.java
@@ -3,7 +3,7 @@ package com.threestar.trainus.domain.coupon.admin.dto;
 import java.util.List;
 
 public record CouponListResponseDto(
-	Long totalCount,
+	Integer totalCount,
 	List<CouponListItemDto> couponList
 ) {
 }
